Comment out all traces of legacy keepassdb module
The new module to be imported is pykeepass, as the keepassdb module
imported from the original script is not available anymore.
An unused import has also been removed.

This is paving the way to a new version using the pykeepass module,
which is available in Manjaro without needing to resort to AUR or
pip, indicating it has been tested and approved by the community
up to a reasonable standard and that it should not disappear into
thin air like the keepassdb module that was initially used.